Fix thinning, assuming hourly backups but not on the hour
is-hourly-backup fails to recognize hourly backups if the cron job doing the
backups runs on a minute other than 0. For example, if we run backups on 5
minutes after the hour every hour, the thinning algorithm will keep 16
sub-hourly backups and never promote any.

Since we don't do any sub-hourly backups, the easiest fix is to assume all
backups are hourly. This also allows us to reduce the number of backups if we
want a month's worth from 70 (16 + 24 + 30) to 54 (24 + 30).